[#83936406] Tweet embed crashes with extra parameters
When fetching via oembed and jsonp, we never recieve errors when the
site returns 404 or 500. Only a success is certain. Therefore, a timeout
of 5 seconds is introduced and stuff that is not successful after that
time is considered failed. Broken tweets and soundcloud embeds can be
removed from the editor that way.

This also fixes the addition of tweets w/ query strings.